+++ src/Support/UpdateMeta/ValueObjects/ChangeMeta.php
@@ -2,12 +2,14 @@
 
 namespace DarkPeople\TelegramBot\Support\UpdateMeta\ValueObjects;
 
+use Telegram\Bot\Objects\ChatMember;
+
 final class ChangeMeta
 {
     public function __construct(
-        public readonly string $kind, // e.g. chat_member
-        public readonly ?array $before,
-        public readonly ?array $after,
+        public readonly ?string $kind = "unknown", // e.g. chat_member|unknown
+        public readonly ?ChatMember $before = null,
+        public readonly ?ChatMember $after = null,
     ) {}
 
     public function hasBeforeAfter(): bool