+++ src/Support/UpdateMeta/ValueObjects/TargetMeta.php
@@ -2,21 +2,30 @@
 
 namespace DarkPeople\TelegramBot\Support\UpdateMeta\ValueObjects;
 
-final class TargetMeta
-{
-    public function __construct(
-        public readonly ?int $id,
-        public readonly bool $is_bot,
-        public readonly ?string $first_name,
-        public readonly ?string $last_name,
-        public readonly ?string $username,
-        public readonly bool $is_premium = false,
-        public readonly array $raw = [],
-        public readonly array $all = [],
-    ) {}
+use Telegram\Bot\Objects\User;
 
-    public function first(int $index = 0) {
-        return $this->all[$index];
+final class TargetMeta extends User
+{
+    public readonly ?string $role; // creator|administrator|member|restricted|left|kicked|unknown (chat role)
+    public readonly array $all;
+
+    public static function fromUser(array $user, ?string $role = null) : self {
+        $target = new self();
+        $target->role = $role;
+        $target->all = $user;
+        
+        $first = $user[0];
+        $target->id = $first->id;
+        $target->isBot = $first->isBot;
+        $target->firstName = $first->firstName;
+        $target->lastName = $first->lastName;
+        $target->username = $first->username;
+        $target->languageCode = $first->languageCode;
+        $target->canJoinGroups = $first->canJoinGroups;
+        $target->canReadAllGroupMessages = $first->canReadAllGroupMessages;
+        $target->supportsInlineQueries = $first->supportsInlineQueries;
+
+        return $target;
     }
 
 }